core: move base functions from "homescreen" app to "base"
