+++ core/SConscript.unix
@@ -454,6 +454,7 @@ if FROZEN:
         ])
     )
 
+    SOURCE_PY.append(SOURCE_PY_DIR + 'apps/base.py')
     SOURCE_PY.extend(Glob(SOURCE_PY_DIR + 'apps/common/*.py'))
     SOURCE_PY.extend(Glob(SOURCE_PY_DIR + 'apps/common/*/*.py'))
     SOURCE_PY.extend(Glob(SOURCE_PY_DIR + 'apps/debug/*.py'))
